+++ sdk/metricsadvisor/azure-ai-metricsadvisor/src/test/java/com/azure/ai/metricsadvisor/DetectionConfigurationTestBase.java
@@ -45,21 +45,21 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
         final String detectionConfigName = "testdetectionconfig" + UUID.randomUUID();
         final MetricWholeSeriesDetectionCondition wholeSeriesCondition = new MetricWholeSeriesDetectionCondition()
             .setConditionOperator(DetectionConditionOperator.OR)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(50)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(13).setMinRatio(50)))
-            .setHardThresholdCondition(new HardThresholdCondition()
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                50,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(13, 50)))
+            .setHardThresholdCondition(new HardThresholdCondition(
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(5, 5))
                 .setLowerBound(0.0)
-                .setUpperBound(100.0)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(5).setMinRatio(5)))
-            .setChangeThresholdCondition(new ChangeThresholdCondition()
-                .setChangePercentage(50)
-                .setShiftPoint(30)
-                .setWithinRange(true)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(2).setMinRatio(2)));
+                .setUpperBound(100.0))
+            .setChangeThresholdCondition(new ChangeThresholdCondition(
+                50,
+                30,
+                true,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(2, 2)));
 
         final AnomalyDetectionConfiguration detectionConfiguration
             = new AnomalyDetectionConfiguration(detectionConfigName)
@@ -144,21 +144,21 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
         final String detectionConfigName = "testdetectionconfig" + UUID.randomUUID();
         final MetricWholeSeriesDetectionCondition wholeSeriesCondition = new MetricWholeSeriesDetectionCondition()
             .setConditionOperator(DetectionConditionOperator.AND)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(50)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(13).setMinRatio(50)))
-            .setHardThresholdCondition(new HardThresholdCondition()
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                50,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(13, 50)))
+            .setHardThresholdCondition(new HardThresholdCondition(
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(5, 5))
                 .setLowerBound(0.0)
-                .setUpperBound(100.0)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(5).setMinRatio(5)))
-            .setChangeThresholdCondition(new ChangeThresholdCondition()
-                .setChangePercentage(50)
-                .setShiftPoint(30)
-                .setWithinRange(true)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(2).setMinRatio(2)));
+                .setUpperBound(100.0))
+            .setChangeThresholdCondition(new ChangeThresholdCondition(
+                50,
+                30,
+                true,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(2, 2)));
 
         final DimensionKey seriesKey = new DimensionKey()
             .put("city", "Shenzhen")
@@ -166,20 +166,20 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
 
         final MetricSingleSeriesDetectionCondition seriesCondition
             = new MetricSingleSeriesDetectionCondition(seriesKey)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)));
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)));
 
         final DimensionKey seriesGroupKey = new DimensionKey()
             .put("city", "Sao Paulo");
 
         final MetricSeriesGroupDetectionCondition seriesGroupCondition
             = new MetricSeriesGroupDetectionCondition(seriesGroupKey)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)));
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)));
 
         final AnomalyDetectionConfiguration detectionConfiguration
             = new AnomalyDetectionConfiguration(detectionConfigName)
@@ -321,21 +321,21 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
         final String detectionConfigName = "testdetectionconfig" + UUID.randomUUID();
         final MetricWholeSeriesDetectionCondition wholeSeriesCondition = new MetricWholeSeriesDetectionCondition()
             .setConditionOperator(DetectionConditionOperator.AND)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(50)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(13).setMinRatio(50)))
-            .setHardThresholdCondition(new HardThresholdCondition()
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                50,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(13, 50)))
+            .setHardThresholdCondition(new HardThresholdCondition(
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(5, 5))
                 .setLowerBound(0.0)
-                .setUpperBound(100.0)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(5).setMinRatio(5)))
-            .setChangeThresholdCondition(new ChangeThresholdCondition()
-                .setChangePercentage(50)
-                .setShiftPoint(30)
-                .setWithinRange(true)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(2).setMinRatio(2)));
+                .setUpperBound(100.0))
+            .setChangeThresholdCondition(new ChangeThresholdCondition(
+                50,
+                30,
+                true,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(2, 2)));
 
         final DimensionKey seriesKey1 = new DimensionKey()
             .put("city", "Shenzhen")
@@ -344,21 +344,21 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
         final MetricSingleSeriesDetectionCondition seriesCondition1
             = new MetricSingleSeriesDetectionCondition(seriesKey1)
             .setConditionOperator(DetectionConditionOperator.OR)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)))
-            .setHardThresholdCondition(new HardThresholdCondition()
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)))
+            .setHardThresholdCondition(new HardThresholdCondition(
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(5, 5))
                 .setLowerBound(0.0)
-                .setUpperBound(100.0)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(5).setMinRatio(5)))
-            .setChangeThresholdCondition(new ChangeThresholdCondition()
-                .setChangePercentage(50)
-                .setShiftPoint(30)
-                .setWithinRange(true)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(2).setMinRatio(2)));
+                .setUpperBound(100.0))
+            .setChangeThresholdCondition(new ChangeThresholdCondition(
+                50,
+                30,
+                true,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(2, 2)));
 
         final DimensionKey seriesKey2 = new DimensionKey()
             .put("city", "Osaka")
@@ -366,10 +366,10 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
 
         final MetricSingleSeriesDetectionCondition seriesCondition2
             = new MetricSingleSeriesDetectionCondition(seriesKey2)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)));
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)));
 
         final DimensionKey seriesGroupKey1 = new DimensionKey()
             .put("city", "Sao Paulo");
@@ -377,31 +377,31 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
         final MetricSeriesGroupDetectionCondition seriesGroupCondition1
             = new MetricSeriesGroupDetectionCondition(seriesGroupKey1)
             .setConditionOperator(DetectionConditionOperator.OR)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)))
-            .setHardThresholdCondition(new HardThresholdCondition()
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)))
+            .setHardThresholdCondition(new HardThresholdCondition(
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(5, 5))
                 .setLowerBound(0.0)
-                .setUpperBound(100.0)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(5).setMinRatio(5)))
-            .setChangeThresholdCondition(new ChangeThresholdCondition()
-                .setChangePercentage(50)
-                .setShiftPoint(30)
-                .setWithinRange(true)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(2).setMinRatio(2)));
+                .setUpperBound(100.0))
+            .setChangeThresholdCondition(new ChangeThresholdCondition(
+                50,
+                30,
+                true,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(2, 2)));
 
         final DimensionKey seriesGroupKey2 = new DimensionKey()
             .put("city", "Seoul");
 
         final MetricSeriesGroupDetectionCondition seriesGroupCondition2
             = new MetricSeriesGroupDetectionCondition(seriesGroupKey2)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)));
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)));
 
         final AnomalyDetectionConfiguration detectionConfiguration
             = new AnomalyDetectionConfiguration(detectionConfigName)
@@ -689,21 +689,21 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
         final String detectionConfigName = "testdetectionconfig" + UUID.randomUUID();
         final MetricWholeSeriesDetectionCondition wholeSeriesCondition = new MetricWholeSeriesDetectionCondition()
             .setConditionOperator(DetectionConditionOperator.AND)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(50)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(13).setMinRatio(50)))
-            .setHardThresholdCondition(new HardThresholdCondition()
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                50,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(13, 50)))
+            .setHardThresholdCondition(new HardThresholdCondition(
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(5, 5))
                 .setLowerBound(0.0)
-                .setUpperBound(100.0)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(5).setMinRatio(5)))
-            .setChangeThresholdCondition(new ChangeThresholdCondition()
-                .setChangePercentage(50)
-                .setShiftPoint(30)
-                .setWithinRange(true)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(2).setMinRatio(2)));
+                .setUpperBound(100.0))
+            .setChangeThresholdCondition(new ChangeThresholdCondition(
+                50,
+                30,
+                true,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(2, 2)));
 
         final DimensionKey seriesKey1 = new DimensionKey()
             .put("city", "Shenzhen")
@@ -712,21 +712,21 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
         final MetricSingleSeriesDetectionCondition seriesCondition1
             = new MetricSingleSeriesDetectionCondition(seriesKey1)
             .setConditionOperator(DetectionConditionOperator.OR)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)))
-            .setHardThresholdCondition(new HardThresholdCondition()
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)))
+            .setHardThresholdCondition(new HardThresholdCondition(
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(5, 5))
                 .setLowerBound(0.0)
-                .setUpperBound(100.0)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(5).setMinRatio(5)))
-            .setChangeThresholdCondition(new ChangeThresholdCondition()
-                .setChangePercentage(50)
-                .setShiftPoint(30)
-                .setWithinRange(true)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(2).setMinRatio(2)));
+                .setUpperBound(100.0))
+            .setChangeThresholdCondition(new ChangeThresholdCondition(
+                50,
+                30,
+                true,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(2, 2)));
 
         final DimensionKey seriesKey2 = new DimensionKey()
             .put("city", "Osaka")
@@ -734,10 +734,10 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
 
         final MetricSingleSeriesDetectionCondition seriesCondition2
             = new MetricSingleSeriesDetectionCondition(seriesKey2)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)));
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)));
 
         final DimensionKey seriesGroupKey1 = new DimensionKey()
             .put("city", "Sao Paulo");
@@ -745,21 +745,21 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
         final MetricSeriesGroupDetectionCondition seriesGroupCondition1
             = new MetricSeriesGroupDetectionCondition(seriesGroupKey1)
             .setConditionOperator(DetectionConditionOperator.OR)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)))
-            .setHardThresholdCondition(new HardThresholdCondition()
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)))
+            .setHardThresholdCondition(new HardThresholdCondition(
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(5, 5))
                 .setLowerBound(0.0)
-                .setUpperBound(100.0)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(5).setMinRatio(5)))
-            .setChangeThresholdCondition(new ChangeThresholdCondition()
-                .setChangePercentage(50)
-                .setShiftPoint(30)
-                .setWithinRange(true)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(2).setMinRatio(2)));
+                .setUpperBound(100.0))
+            .setChangeThresholdCondition(new ChangeThresholdCondition(
+                50,
+                30,
+                true,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(2, 2)));
 
         final AnomalyDetectionConfiguration detectionConfiguration
             = new AnomalyDetectionConfiguration(detectionConfigName)
@@ -777,10 +777,10 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
 
         final MetricSeriesGroupDetectionCondition seriesGroupConditionToAddOnUpdate
             = new MetricSeriesGroupDetectionCondition(seriesGroupKeyToAddOnUpdate)
-            .setSmartDetectionCondition(new SmartDetectionCondition()
-                .setSensitivity(63)
-                .setAnomalyDetectorDirection(AnomalyDetectorDirection.BOTH)
-                .setSuppressCondition(new SuppressCondition().setMinNumber(1).setMinRatio(100)));
+            .setSmartDetectionCondition(new SmartDetectionCondition(
+                63,
+                AnomalyDetectorDirection.BOTH,
+                new SuppressCondition(1, 100)));
     }
 
     protected DataFeed createDataFeed(HttpClient httpClient,
@@ -792,11 +792,11 @@ public abstract class DetectionConfigurationTestBase extends MetricsAdvisorAdmin
             SQL_SERVER_CONNECTION_STRING,
             TEMPLATE_QUERY));
         dataFeed.setSchema(new DataFeedSchema(Arrays.asList(
-            new DataFeedMetric().setName("cost").setDisplayName("cost"),
-            new DataFeedMetric().setName("revenue").setDisplayName("revenue")))
+            new DataFeedMetric("cost").setDisplayName("cost"),
+            new DataFeedMetric("revenue").setDisplayName("revenue")))
             .setDimensions(Arrays.asList(
-                new DataFeedDimension().setName("city").setDisplayName("city"),
-                new DataFeedDimension().setName("category").setDisplayName("category"))))
+                new DataFeedDimension("city").setDisplayName("city"),
+                new DataFeedDimension("category").setDisplayName("category"))))
             .setName("java_data_feed_for_detection" + UUID.randomUUID())
             .setGranularity(new DataFeedGranularity().setGranularityType(DataFeedGranularityType.DAILY))
             .setIngestionSettings(new DataFeedIngestionSettings(INGESTION_START_TIME));