+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/aad/webapp/AADOAuth2AuthorizationCodeGrantRequestEntityConverter.java
@@ -3,8 +3,8 @@
 
 package com.azure.spring.aad.webapp;
 
-import com.azure.spring.common.AbstractOAuth2AuthorizationCodeGrantRequestEntityConverter;
 import com.azure.spring.core.ApplicationId;
+import com.azure.spring.aad.AbstractOAuth2AuthorizationCodeGrantRequestEntityConverter;
 import org.springframework.http.RequestEntity;
 import org.springframework.security.oauth2.client.endpoint.OAuth2AuthorizationCodeGrantRequest;
 import org.springframework.util.LinkedMultiValueMap;