+++ sdk/spring/azure-spring-boot-starter-keyvault-certificates/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 3.1.0-beta.1 (Unreleased)
+
+
 ## 3.0.0 (2021-06-23)
 ### New Features
 - Load JRE key store certificates to AzureKeyVault key store. ([#21845](https://github.com/Azure/azure-sdk-for-java/pull/21845))