+++ sdk/spring/azure-spring-boot-starter-keyvault-secrets/CHANGELOG.md
@@ -1,11 +1,25 @@
 # Release History
 
 ## 3.7.0-beta.1 (Unreleased)
-
+### New Features
+- Support unified properties as below when the associated "azure.keyvault.*" properties are not configured.([#22396](https://github.com/Azure/azure-sdk-for-java/issues/22396))
+    ```
+    spring:
+      cloud:
+        azure:
+          client-id:
+          client-secret:
+          certificate-path:
+          tenant-id:
+    ```
+### Breaking Changes
+- Deprecate property of "azure.keyvault.client-key", and provide ""azure.keyvault.client-secret" instead.
 
 ## 3.6.0 (2021-06-23)
 ### Breaking Changes
-- Deprecated `allowTelemetry` configuration item.
+
+### Deprecations
+- Deprecate `allowTelemetry` configuration item.
 
 ## 3.5.0 (2021-05-24)
 ### New Features