+++ sdk/spring/azure-spring-boot-starter/CHANGELOG.md
@@ -1,7 +1,20 @@
 # Release History
 
 ## 3.7.0-beta.1 (Unreleased)
-
+### New Features
+- Support unified properties as below when the service specific properties are not configured.([#22396](https://github.com/Azure/azure-sdk-for-java/issues/22396))
+    ```
+    spring:
+      cloud:
+        azure:
+          authority-host:
+          client-id:
+          client-secret:
+          certificate-path:
+          msi-enabled:
+          tenant-id:
+          environment:
+    ```
 
 ## 3.6.0 (2021-06-23)
 ### New Features